backend/requirements.txt
Flask>=2.0.0
flask-cors>=4.0.0
boto3>=1.26.0
botocore>=1.29.0
Pillow>=9.0.0
numpy>=1.21.0
opencv-python>=4.5.0
redis>=4.0.0
bcrypt>=4.0.1
PyJWT>=2.8.0
pymongo>=4.0.0
python-dotenv>=1.0.0
supabase>=2.19.0
